Update README.md
For the following program I was tasked with writing code that would utilize a fruitful function. First and foremost, a fruitful function is defined as "A function that yields a return value". In order to emphasize the changing output of fruitful functions, I decided to express it visually using the turtle module. The turtle module is a drawing module that can be imported into the code and manipulated using command codes.  
The code I wrote is based on the recurrences of circles and allows the user to manipulate the size and the complexity of the picture. The choices are decided by giving the user set options that are controlled in if and elif statements and stored. The answers given such as "moderate" or "intense" correspond to set numeric values that allow for the function to continue. These stored decisions are then later used to give values to variables assigned in functions that I had built and defined, allowing them to run.